+++ src/components/header/Header.module.scss
@@ -31,8 +31,8 @@
       }
 
       .addRestaurantButton {
-        background-color: $titleColor !important;
-        color: white !important;
+        background-color: $titleColor;
+        color: white;
       }
     }
   }
