+++ src/lib/RestaurantServices.js
@@ -9,22 +9,24 @@ const SERVER_CALL = axios.create({
 export const fetchRestaurants = async () =>
   (await SERVER_CALL.get("/restaurants")).data;
 
+const formatRestaurantObject = data => ({
+  restaurant: {
+    name: data.name,
+    cuisine: data.cuisine,
+    is_10_bis: !!data.is10Bis,
+    address: data.address,
+    max_delivery_time_in_minutes:
+    data.maxDeliveryTimeInMinutes
+  }
+});
+
 export const createRestaurant = async restaurant =>
   (await SERVER_CALL.post(
     "/restaurants",
     restaurant,
     {
       transformRequest: [
-        data => ({
-          restaurant: {
-            name: data.name,
-            cuisine: data.cuisine,
-            is_10_bis: !!data.is10Bis,
-            address: data.address,
-            max_delivery_time_in_minutes:
-              data.maxDeliveryTimeInMinutes
-          }
-        }),
+        formatRestaurantObject,
         ...axios.defaults.transformRequest
       ]
     }
