+++ src/selectors/FiltersSelectors.js
@@ -1,0 +1,8 @@
+import { createSelector } from "reselect";
+
+const getFiltersValues = state => state.restaurantReducer.filters.values;
+
+export const getRestaurantsFilters = createSelector(
+  [getFiltersValues],
+  filtersValues => filtersValues
+);
